Pg_Stat_Guard/requirements.txt
azure-functions
tdigest
scipy
numpy